20250104 csy

+++ yumyum/src/main/java/org/koreait/yumyum/dto/rating/response/RatingMonthResponseDto.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class RatingMonthResponseDto {
     private String reviewMonth;
-    private int avgRating;
+    private Double avgRating;
     private int reviewMonthCount;
 }
